requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
langgraph
python-dotenv